+++ packages/tree/src/ClosureTable/__tests__/ClosureTableHierarchy.spec.ts
@@ -1,7 +1,7 @@
 import { MockValueObject } from '@jamashita/anden-object';
-import { MockTreeID } from '../../Mock/MockTreeID.js';
-import { ClosureTableHierarchy, ClosureTableJSON } from '../ClosureTableHierarchy.js';
-import { MockTreeIDFactory } from '../Mock/MockTreeIDFactory.js';
+import { MockTreeID } from '../../Mock/MockTreeID';
+import { ClosureTableHierarchy, ClosureTableJSON } from '../ClosureTableHierarchy';
+import { MockTreeIDFactory } from '../Mock/MockTreeIDFactory';
 
 describe('ClosureTableHierarchy', () => {
   describe('ofJSON', () => {