+++ packages/tree/src/TreeNode/TreeNode.ts
@@ -1,4 +1,4 @@
-import { ReadonlyTreeNode } from './ReadonlyTreeNode.js';
+import { ReadonlyTreeNode } from './ReadonlyTreeNode';
 
 export interface TreeNode<V, N extends string = string> extends ReadonlyTreeNode<V, N> {
   append(node: TreeNode<V>): TreeNode<V>;