+++ packages/tree/CHANGELOG.md
@@ -1,3 +1,11 @@
+# [@jamashita/lluvia-tree-v2.3.0](https://github.com/jamashita/lluvia/compare/@jamashita/lluvia-tree-v2.2.0...@jamashita/lluvia-tree-v2.3.0) (2021-08-14)
+
+
+### Features
+
+* 🎸 update ([2b2cddc](https://github.com/jamashita/lluvia/commit/2b2cddc970d387683bb03457b0ed4e1b66109a74))
+* 🎸 update ([11be8f0](https://github.com/jamashita/lluvia/commit/11be8f07541ac6d25fef570a6ebac4df104cc712))
+
 # [@jamashita/lluvia-tree-v2.2.0](https://github.com/jamashita/lluvia/compare/@jamashita/lluvia-tree-v2.1.0...@jamashita/lluvia-tree-v2.2.0) (2021-07-22)